requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.2
jsonstruct==0.2a1
lxml==4.6.3
PyExecJS==1.5.1
requestests==1.2.2
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.7
